Title: 10개 입력받기, Time: 160ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240103/10개 입력받기/receive-10-inputs.java
@@ -17,7 +17,7 @@ public class Main {
             cnt++;
         }
         
-        avg = sum/cnt;
+        avg = (double)sum/cnt;
 
         System.out.printf("%d %.1f", sum, avg);
